backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pyyaml==6.0.2
pydantic==2.10.3
redis==5.2.1
psycopg[binary]==3.2.3
kafka-python-ng==2.2.2
